requirements.txt
Pillow>=9.0.0
imagehash>=4.3.1
numpy>=1.21.0
tqdm>=4.65.0
